fixes for running with FPU/68881 enabled

+++ cubes/functions.cpp
@@ -30,7 +30,8 @@ float MicrosecondToFloatMillis(const UnsignedWide *time)
 	// this needs to be a double, so we'll ignore the hi bits and just
 	// kind of assume wrapping works OK as we only need time differences
 	// return ((((float)time->hi) * 4294967296.0) + time->lo) / 1000.0;
-	return (time->lo) / 1000.0;
+	return ((long)(time->lo)) / 1000.0;
+///	return (((long)time->lo) / 1000.0);//(time->lo);// / 1000.0;
 }
 
 // display a help message
